requirements.txt
opencv-python==4.8.1.78
opencv-contrib-python==4.8.1.78
numpy==1.24.3
scikit-image==0.22.0
matplotlib==3.7.2
pandas==2.0.3
Pillow==10.0.0
scipy==1.11.2
cellpose==2.2.3
reportlab==4.0.4